--- cockroachdb/tests/utils.py
@@ -1,12 +1,0 @@
-# (C) Datadog, Inc. 2023-present
-# All rights reserved
-# Licensed under a 3-clause BSD style license (see LICENSE)
-import os
-
-from datadog_checks.dev import get_here
-
-HERE = get_here()
-
-
-def get_fixture_path(filename):
-    return os.path.join(HERE, 'fixtures', filename)
